🔨 refactor duplicated functions
